ask before each cleaning step

+++ resync.py
@@ -60,13 +60,18 @@ parser.add_argument('--transfer-dir', metavar='<directory name>', dest='prepdir'
                     help='custom directory to render files to-be-upload')
 parser.add_argument('--debug', dest='debug', action='store_true', default=False,
                     help="Render documents, but don't copy to remarkable.")
+parser.add_argument('-y', '--yes', dest='yes', action='store_true', default=False,
+                    help="Do not ask for deletion.")
 
 parser.add_argument('mode', type=str, choices=["push","pull","backup","+","-","clean"],
                     help=("Specifies the transfer mode."
                           "\n  push, +: push documents from this machine to the reMarkable"
                           "\n  pull, -: pull documents from the reMarkable to this machine"
                           "\n  backup:  pull all files from the remarkable to this machine (excludes still apply)"
-                          "\n  clean:   clear the files in a trash, and remove the orphaned files"))
+                          "\n  clean:   performs a number of cleaning operations."
+                          "\n           * clear the files in Trash"
+                          "\n           * remove the orphaned files"
+                          "\n           We ask before performing each step unless -y|--yes."))
 
 parser.add_argument('documents', metavar='documents', type=str, nargs='*',
                     help='Documents and folders to be pushed to the reMarkable')
@@ -263,6 +268,14 @@ def curb_tree(node, excludelist):
     node.children = uncurbed_children
 
     return False
+
+
+
+def ask(msg):
+    if args.yes:
+        return True
+    else:
+        return input(msg+" [Enter,y,Y / n]") in ['', 'y', 'Y']
 
 
 #################################
@@ -710,6 +723,7 @@ def pull_from_remarkable():
 
 
 def cleanup_deleted():
+    """remove trash files"""
 
     deleted_uuids = []
     limit = 10
@@ -720,8 +734,7 @@ def cleanup_deleted():
     if len(deleted_uuids) == 0:
         print('No deleted files found.')
     else:
-        decision = input(f'Clean up {len(deleted_uuids)} deleted files? [Y/n]')
-        if decision in ['', 'y', 'Y']:
+        if ask(f'Clean up {len(deleted_uuids)} deleted files?'):
             for u in deleted_uuids:
                 ssh("rm -r ~/.local/share/remarkable/xochitl/{u}*", dry=args.dryrun)
 
@@ -729,10 +742,12 @@ def cleanup_deleted():
 
 
 def cleanup_orphaned():
+    """remove pdfs that lack metadata"""
     if args.dryrun:
         ssh('"for f in $(ls -1 ~/.local/share/remarkable/xochitl) ; do stem=${$(basename $f)%%.*}; if ! [ -e $stem.metadata ] ; then echo rm $stem.* ; fi ; done"')
     else:
-        ssh('"for f in $(ls -1 ~/.local/share/remarkable/xochitl) ; do stem=${$(basename $f)%%.*}; if ! [ -e $stem.metadata ] ; then rm $stem.* ; fi ; done"')
+        if ask(f'Clean up orphaned pdfs?'):
+            ssh('"for f in $(ls -1 ~/.local/share/remarkable/xochitl) ; do stem=${$(basename $f)%%.*}; if ! [ -e $stem.metadata ] ; then rm $stem.* ; fi ; done"')
 
 
 ssh_connection = None
